projects/cybersecurity-homelab/requirements.txt
requests==2.32.5
beautifulsoup4==4.14.3
scapy==2.7.0
pycryptodome==3.23.0
paramiko==4.0.0
pandas==2.3.3
numpy==2.4.1
python-dotenv==1.2.1
pyyaml==6.0.3
colorama==0.4.6
rich==14.2.0
pytest==9.0.2
black==25.12.0
flake8==7.3.0
